+++ core/hooks.ts
@@ -1,9 +1,9 @@
-import { createListenerHook, createMutatorHook, createMutatorHookAsync } from '@vue-storefront/core/lib/hooks'
+import { createListenerHook, createMutatorHook, createAsyncMutatorHook } from '@vue-storefront/core/lib/hooks'
 
 const {
   hook: beforeStoreViewChangedHook,
   executor: beforeStoreViewChangedExecutor
-} = createMutatorHookAsync<any, any>()
+} = createAsyncMutatorHook<any, any>()
 
 const {
   hook: afterStoreViewChangedHook,