+++ CHANGELOG.md
@@ -26,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Unit Tests for Core filters - @lukaszjedrasik (#5036)
 - Added information about i18n & Unit tests in modules to docs - @lukaszjedrasik (#4991)
 - Added unit test for async-data-loader.ts - @lukaszjedrasik [(#5051)](https://github.com/vuestorefront/vue-storefront/issues/5051)
-- Added `createMutatorHookAsync` & unit tests for hooks - @lukaszjedrasik [(#4940)](https://github.com/vuestorefront/vue-storefront/issues/4940)
+- Added `createAsyncMutatorHook` & unit tests for hooks - @lukaszjedrasik [(#4940)](https://github.com/vuestorefront/vue-storefront/issues/4940)
 
 ### Fixed